Backend/requirements.txt
fastapi
uvicorn[standard]
pillow
torch
torchvision
openai
paho-mqtt
